feat: Enhance lookboard sharing and display
Adds `sharedByUsername`, `clientName` to `SharedLookboardInstance` type for better tracking of shared lookboards.

Introduces a `ChevronDownIcon` component to fix an import error.

The `Lookbook` and `CreatorStudio` pages are updated to accept and display `sharedInstances` of lookboards.

`LookboardsList` component is refactored to display individual shared links for a lookboard, including general link and client-specific links. Includes functionality to copy links and shows creator/username information.

The `ViewLookboardPage` is updated to handle video assets in addition to images for looks when displaying feedback and comments.